Grapher: Word related charts

+++ src/main/java/gr/auth/sam/tredingfeelings/IPlotter.java
@@ -3,6 +3,7 @@ package gr.auth.sam.tredingfeelings;
 
 import java.util.List;
 
+
 /**
  * Creates basic plots that can be shown with swing or be exported into a PNG
  * images.
@@ -14,7 +15,7 @@ public interface IPlotter {
             String xname, List<String> xvalues, String yname, List<Integer> yvalues);
 
     void createZipfChart(String title, int width, int height, //
-            String xname, List<Integer> xvalues, String yname, List<Integer> yvalues);
+            String xname, List<Float> xvalues, String yname, List<Float> yvalues);
 
     void createCumulativeChart(String title, int width, int height, //
             String xname, List<Integer> xvalues, String yname, List<Integer> yvalues);
@@ -30,5 +31,5 @@ public interface IPlotter {
      * @param path the path of the exported image
      */
     void exportChart(String path);
-    
+
 }
